+++ plugins/scaffolder-backend/src/scaffolder/tasks/NunjucksWorkflowRunner.ts
@@ -55,7 +55,7 @@ import { actionExecutePermission } from '@backstage/plugin-scaffolder-common/alp
 import { TaskRecovery } from '@backstage/plugin-scaffolder-common';
 import { PermissionsService } from '@backstage/backend-plugin-api';
 import { loggerToWinstonLogger } from '@backstage/backend-common';
-import { WinstonLogger } from './logger';
+import { BackstageLoggerTransport, WinstonLogger } from './logger';
 import { restoreWorkspace } from './serializer';
 
 type NunjucksWorkflowRunnerOptions = {
@@ -99,9 +99,11 @@ const isValidTaskSpec = (taskSpec: TaskSpec): taskSpec is TaskSpecV1beta3 => {
 const createStepLogger = ({
   task,
   step,
+  rootLogger,
 }: {
   task: TaskContext;
   step: TaskStep;
+  rootLogger: winston.Logger;
 }) => {
   const stepLogStream = new PassThrough();
   stepLogStream.on('data', async data => {
@@ -118,8 +120,8 @@ const createStepLogger = ({
       winston.format.simple(),
     ),
     transports: [
-      new winston.transports.Console(),
       new winston.transports.Stream({ stream: stepLogStream }),
+      new BackstageLoggerTransport(rootLogger),
     ],
   });
 
@@ -263,7 +265,11 @@ export class NunjucksWorkflowRunner implements WorkflowRunner {
 
       const action: TemplateAction<JsonObject> =
         this.options.actionRegistry.get(step.action);
-      const { taskLogger, streamLogger } = createStepLogger({ task, step });
+      const { taskLogger, streamLogger } = createStepLogger({
+        task,
+        step,
+        rootLogger: this.options.logger,
+      });
 
       if (task.isDryRun) {
         const redactedSecrets = Object.fromEntries(