+++ plugins/proxy-backend/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-proxy-backend
 
+## 0.4.16-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.21.8-next.0
+  - @backstage/backend-plugin-api@0.6.18-next.0
+  - @backstage/config@1.2.0
+
 ## 0.4.15
 
 ### Patch Changes